Feat: Implement CSV Import and Finalize ACL
This commit introduces the "Import from CSV" feature and completes the implementation of the ACL security layer.

CSV Import Feature:
- Adds a new "Import from File" section to the administrator backend.
- Allows users to upload a CSV file and create an import job.
- Provides a two-step process with a column mapping screen and data preview.
- Implements the import logic, including de-duplication based on email and correct handling of lead group associations via the LeadModel.

ACL and Refactoring:
- Implements ACL checks in all administrator Views to conditionally show/hide toolbar buttons based on user permissions.
- Refactors all list controllers to extend AdminController, cleaning up the code and leveraging framework-native, ACL-aware actions.

+++ com_crm_joomla/administrator/src/View/Campanhas/HtmlView.php
@@ -12,7 +12,6 @@ namespace Joomla\Component\Crm\Administrator\View\Campanhas;
 use Joomla\CMS\MVC\View\ListView;
 use Joomla\CMS\Toolbar\ToolbarHelper;
 use Joomla\CMS\Language\Text;
-use Joomla\CMS\Factory;
 
 /**
  * Campanhas View
@@ -31,7 +30,7 @@ class HtmlView extends ListView
         $this->activeFilters = $this->get('ActiveFilters');
 
         if (count($errors = $this->get('Errors'))) {
-            Factory::getApplication()->enqueueMessage(implode("\n", $errors), 'error');
+            $this->getApplication()->enqueueMessage(implode("\n", $errors), 'error');
             return;
         }
 
@@ -46,10 +45,24 @@ class HtmlView extends ListView
     protected function addToolbar()
     {
         ToolbarHelper::title(Text::_('COM_CRM_CAMPANHAS_VIEW_DEFAULT_TITLE'));
-        ToolbarHelper::addNew('campanha.add');
-        ToolbarHelper::editList('campanha.edit');
-        ToolbarHelper::deleteList(Text::_('COM_CRM_CONFIRM_DELETE'), 'campanhas.delete');
-        ToolbarHelper::publish('campanhas.publish', 'JTOOLBAR_PUBLISH', true);
-        ToolbarHelper::unpublish('campanhas.unpublish', 'JTOOLBAR_UNPUBLISH', true);
+
+        $user = $this->getApplication()->getIdentity();
+
+        if ($user->authorise('core.create', 'com_crm')) {
+            ToolbarHelper::addNew('campanha.add');
+        }
+
+        if ($user->authorise('core.edit', 'com_crm')) {
+            ToolbarHelper::editList('campanha.edit');
+        }
+
+        if ($user->authorise('core.delete', 'com_crm')) {
+            ToolbarHelper::deleteList(Text::_('COM_CRM_CONFIRM_DELETE'), 'campanhas.delete');
+        }
+
+        if ($user->authorise('core.edit.state', 'com_crm')) {
+            ToolbarHelper::publish('campanhas.publish', 'JTOOLBAR_PUBLISH', true);
+            ToolbarHelper::unpublish('campanhas.unpublish', 'JTOOLBAR_UNPUBLISH', true);
+        }
     }
 }
